Add ComponentType model
Rather than storing array of possible component types in Component
model, add new model for this. This model then has a name and
description for the type of component, and Component no longer needs a
description (I think) as the ComponentType should adequately describe a
component's purpose.

Having this model will allow us to specify the type of Component a given
Case (to be added) expects, if any.