Generalize Contact to User to also support admins
We need admin users, to log in to the admin dashboard at least and
presumably later to also interact with site contacts to resolve support
cases, so this commit generalizes the Contact model to a User model
which can then be used in both these cases.

Clearance requires us to use one model for all models using
authentication, but this should work well for our use case. We can split
out additional models in future which are conditionally associated with
a User depending on whether they are a contact or an admin, if we later
need additional state stored for either of these 2 user types.