Create rt ticket and save id on Case creation
Every Case must have an associated, unique rt ticket, which we now
create and associate it with the Case when this is created. Rails
validations along with how we create rt tickets should mean that it is
now impossible to create a Case without also creating a new associated
rt ticket.