Unselects the component if it's nolonger valid
In the new case form, which components to show is determined by the
case category and cluster. As the user changes this field, the valid
components will change.

However which component will not change. This means an invalid component
might be selected, but can not be unselected if the component list
is empty.

This commit fixes the issue by unselecting the selected component if
it is hidden by the filters.