Hide the component field if it is empty
In some cases, their will be no valid component for a case. When this
occurs, the component field is hidden.

Note: Their is an edge case where the component field is hidden even
if a component is required. This is because the component field will
be hidden if no valid components are found for the case. This means it
is not possible for the user to create that case. However it might be
better to deal with this server side.