Add support cases archive page
This shows all support cases, unlike the main support cases table which
only includes open ones, and it is linked from that table.

Archived cases are shown as greyed out and non-interactive rows.
Currently this is only client-side, so if they really want to users can
still perform the unavailable actions; since neither of these actions
really changes anything within rt or Supportware from an admin's
perspective this is fine at the moment.