Support generating many components at once for group
By setting the `genders_host_range` on a `ComponentGroup` and then
saving it, many components can be created for a group simultaneously.
This is done by expanding the host range using `nodeattr`, and then
creating associated `Component`s for any expanded names which don't yet
have a `Component` in the group. This works both from the admin
dashboard and when interacting with a `ComponentGroup` object directly.