Pass initial form state as flags and fake decoding
The data to be used by the new case form is now passed in by reading
data attributes containing JSON data and passing this in as flags to the
Elm app. These flags will then be decoded in Elm to get the initial app
Model; this does not yet exist and is currently faked.