Only include needed data when initialize case form
Rather than including all data when initializing new Case form, specify
exactly the data we want to include. Including all the data is wasteful
and could potentially expose things we don't want to to users;
additionally we also want to include fields like the value of
`Component#cluster_id` which are not direct attributes of the model and
so would not be included by just using `to_json`.