Ensure initial component part of initial cluster
Previously we were relying on the first component being part of the
first cluster, when these are loaded, to set the initial form state, but
there is no reason that this should be the case. Now we choose the first
component from just those for the initial cluster, so it should always
now be valid.