Split out Issue model from CaseCategory
We have a requirement that there should be 2 levels for classifying
support issues: the CaseCategory, which holds a group of related Issues;
and the Issue, which is a specific named issue type, one of which should
be associated with every support case and which may or may not require
that an associated component or other data is also included with that
support case.

This commit implements the full refactoring to split the current
CaseCategory model into these new models, for both the Rails server and
the case form Elm app.

Since we did not in the end have the requirement for a CaseCategory (or
after this refactoring an Issue) to be associated with a particular
component type, I have also removed the `component_type` field from
CaseCategory at the same time, and not moved this to Issue.