Display details templates in case form
The details template associated with each Issue is now displayed when
that issue is selected. Any changes to the details field will then be
saved associated with that Issue, so that the details field is updated
as the issue is changed, but if the user returns to an issue where they
had previously entered input then their changes will persist.

Also needed to change the `textarea` to have the text be set via the
`value` attribute rather than a nested `text` element, in order to have
this consistently update when the selected case category/issue is
changed. This change appears to resolve this, in Chrome at least, though
we should probably check this is resolved in other browsers (related
issue: https://github.com/elm-lang/html/issues/52).