Add support_type field to Components and Issues
Issues specify the `support_type` they are for, which can currently be
`advice` or `managed`; the latter will mean this issue will only be able
to be associated with a `managed` Cluster or Component. Individual
Components also now have a `support_type`, which can be `advice`,
`managed`, or `inherit`; the latter will mean the Component will use the
`support_type` that its cluster has set.