Validate that Case Issue valid for relevant support type
A Case can only be opened for an Issue with `managed` `support_type` if
either:

- the Issue does not require that a Component be associated with a Case
opened for this Issue, and the associated Cluster is `managed`;

- the Issue does require that a component be associated with a Case
opened for this Issue, and the associated Component is `managed`.

This commit adds validations for these requirements, with tests to
ensure the validations pass or fail appropriately for different
combinations of the models that can/must be associated with a Case.