Add validation to Case form app
Add validation to Case form app for whether the details field is filled,
as well as whether an Issue, and based off this a CaseCategory, is able
to be selected given the `support_type` for this Issue and the currently
selected `Cluster`.

Displaying of validations is done using Bootstrap's so-called
Server-side validations
(https://getbootstrap.com/docs/4.0/components/forms/#server-side), as
this allows us to simply specify in Elm whether a field is valid or not
by adding a class to it.

The state for the validation status of a field is handled using a new
`FieldValidation` type, with a corresponding module for functions
directly related to this.