Add asset record related models
Every ComponentType has many AssetRecordFieldDefinitions, specifying a
field name and the highest level it can be set at (group-level, which
will be overridable for nodes, or node only).

Every Component or ComponentGroup can have many AssetRecordFields, which
are individual specific values for the AssetRecordFieldDefinitions for
the ComponentType associated with that Component's ComponentGroup or the
ComponentGroup directly.