Add method for whether record field can be set at group-level
Currently, asset record fields can always be set at Component-level
(i.e. the AssetRecordField row is associated with a Component), but they
may only be set at ComponentGroup-level (i.e. the AssetRecordField row
is associated with a ComponentGroup) if the AssetRecordFieldDefinition
is specified with a `group` `level` attribute. This commit adds a method
to determine if this is the case.