Add validation that only 1 of field per asset
Having more than 1 AssetRecordField for a given
AssetRecordFieldDefinition and the same Component or ComponentGroup does
not make sense, as we would not know which would take precedence, and
would not want to display both to users. This commit adds a validation
to prevent this situation from occurring.

Also needed to add a `reload` of a model in a different test to get this
passing for some reason.