Tweak admin interface for various models
Hide various fields we do not want to make editable for different
models. In particular given our data model it often doesn't make much
sense to allow configuring a one-to-many relationship from the many side
of the relationship. Also prevent changing of the Cluster/Site
`canonical_name` field, as doing so could break the look-up of related
files on S3.