Thread through new rt `cc` parameter
This is about the minimal code needed to the current tests to pass with
a new `cc` parameter. We don't actually test the body of the request to
be sent to rt yet, so this parameter is not actually being used yet;
adding a test for this and then having this parameter be used will be
done shortly.