Add Issues to request switching Component status
The Issue to request that a Component be switched from `advice` to
`managed` should only be available when a Component has an `advice`
`support_type`. Since this is a new situation (previously an `advice`
issue would be available for both `managed` and `advice` Components), we
need to introduce a new option for an `Issue`s `support_type`.