Display more informative message for invalid Component
Components can now be unavailable for the selected Issue for a couple of
reasons depending on their SupportType, so the error message shown is
now changed based on this. In order to do this FieldValidation needed to
be changed to allow the error message to be varied based on the
particular item which is invalid, and code using this therefore needed
to be adapted.