Implement custom password reset process
For Alces Flight Center password resets we want:

1. The user requests a password reset.

2. The user receives an email with a (unique, secure) link to reset
their password; their password remains unchanged at this point.

3. If this link is clicked then a new password is generated for the
user, they are signed in, and the new password is emailed to them.

This commit implements this full process, with a feature spec to ensure
it all works together; the standard Clearance password reset process
(which partially handles points (1) and (2) above) has been adapted to
do this.