Fix rails-admin config being inconsistently loaded
A couple of issues appeared to be causing the rails-admin config for
models to be inconsistently loaded. This was particularly noticeable in
production as when developing the config for a particular model this
appeared to be auto-reloaded, but in production configs often seemed to
not be loaded at all. These issues were:

- Defining the config for each model as a module namespaced within that
model, and/or storing these modules within the main `models` directory,
appeared to be the main issue. Moving the config module's to be
namespaced under `AdminConfig` and within the
`models/concerns/admin_config` directory appears to resolve this.

- It is apparently only possible to define a `rails_admin` config block
once for a model, and if this is done multiple times then one will be
silently ignored. This was effecting `Component` and `ComponentGroup` as
they had a config block defined both specific to them and in the shared
`EditableAssetRecordFields`. Changing how this works to only define the
config in one place appears to resolve this.