Disable group fields only settable for Component
Certain asset record fields can only be set for individual Components,
and not at the ComponentGroup level, as it would not make sense to have
one value for these fields shared across all Components in a group. For
example this applies to the serial number or IP fields.

We were enforcing this constraint at the model level, with validations,
but not at the admin interface level; this was leading to errors with no
user-visible explanation shown when values for these fields were entered
at the group level.

This commit adds the ability to disable an asset record field input
using the properties of the definition; this is then used to disable
fields which can only be set for Components when editing a
ComponentGroup, in which case explanatory title text is shown.