Use AR associations when creating seed data
This simplifies the creation of objects in various cases, as don't need
to both create an object and link back to 'parent' object, and instead
can do this in one line. It also makes the 'parent' immediately aware of
the new object, which is useful in having any validations use the latest
data, and avoiding having to reload objects to see latest associated
objects.