Add identifiers to Issues
We need to be able to find and use certain Issues in custom ways, such
as convenient quick Case creation buttons for certain kinds of issues.
The concept of optional, unique identifiers allows us to do this, while
still allowing all Issues to be treated as user-editable data.

This commit also adds a couple of identifiers to the seed data for the
Component support_type switching issues, along with convenience finder
methods for these Issues.