Add possible Case-Service relationship
We want the ability to have Issues that require that a Case have a
Cluster Service associated. The Case-Service relationship should
function similarly to the Case-Component one, hence the use of shared
examples for most of the tests for these relationships.

This commit adds this, along with the various required validations, in
the most straightforward way. No seeded Issues yet have the
`requires_service` flag set, and the validation code may still be
improved in a later commit to reduce the duplication that now exists.

Note that this concept is about Issues and Cases for the management of
Cluster Services themselves. There is also the concept, which no related
code yet exists for, of Issues being related to a particular Service.
For example an Issue related to requesting new user accounts is related
to a Cluster's 'User Management' Service. We may later add functionality
to be able to display Cases for such Issues along with the associated
Service, as well as to prevent creation of Cases for certain Issues
given the support_type of the associated Service.