Generalize association info shown for Case
Previously in the Case table we just showed the Component name, if a
Case was associated with a Component, or nothing if not.

This commit retains showing the Component name while also showing the
Service name if a Case is associated with a Service and no Component, or
indicating if neither of these associations is present. When a Component
or Service is shown, the corresponding page for this is also linked.