Introduce ClusterPart type; refactor to use
This commit introduces a ClusterPart type, to be used for the common
features of Components and Services, and adapts the current Components
field and validation to be generic using this. This should make it
simple to slot in similar behaviour for Services, in a type-safe way and
reusing the existing functions.