Handle single part Case form in generic way
Generalize how we previously handled setting up the 'single component
mode' new Case form to switch based off a more generic singlePart flag
which is now passed into the app, and then decode and use this data to
determine what the initial app state should be. The new 'single service
mode' is not yet being specifically handled; this will be added in a
later commit.