Add rails-data-migrations Gem
So can add new data to be added to the production database while also
having this be seeded in development.

This should work for now, but as production data is changed something
more robust might be needed to ensure consistent environments, such as
importing all production data, or at least the tables we care about
(CaseCategory, Issue, ServiceType etc).