Filter Issues when filter CaseCategorys
Previously, when filtering the CaseCategorys to be shown in the single
Component or Service mode Case form, we would show all Issues in a
CaseCategory if any of them passed the given condition (which is
currently whether the Issue requires a Component/Service).

However, we actually only want to show only those Issues which
themselves pass the condition, as well as the containing CaseCategorys;
this commit does this.