Label interface test so can easily distinguish
We have 2 kinds of tests for the RT interface classes: tests that
certain interface invariants hold to ensure the fake interface is
functioning like the real one, and tests for the individual
implementations to ensure their internals are working as required. This
commit adds a label to the former to distinguish it from the latter.