Allow admins to view all Cases create Site Cases
Because admins are able to access all Sites, we need to provide
Site-namespaced routes for viewing all Cases for a Site and for creating
a Case for a Site, whereas for regular users we just implicitly use the
single Site a user is associated with. The same CasesController still
works in both cases however.

We also need to tweak some links to point to these paths instead when
the current user is an admin.