Adapt Site Cases page for admins
When an admin is viewing the page with all Cases for a Site, we do not
want to show the contact or archive/restore buttons, or indicate whether
a Case has been archived by a Site contact; these features are only
intended to help Site contacts manage their Cases.

All other places the Cases table is shown remain the same for the
moment.