Add CreditCharge forms to manage Cases page
For those Cases which are both chargeable and complete we now show a
small form within the manage Cases page to associate a CreditCharge with
that Case.

If a Case falls into this group and has no CreditCharge already, then
its row is highlighted to indicate that this is needed.

If a Case does not fall into this group, and so cannot (at least yet)
have a CreditCharge associated, a brief indication of why is given.