Show credits to contacts in Cases table
We want to indicate the usage of credits to Site contacts, this is a
first pass at doing this.

If a User is a Site contact we show credit fields similarly to admins,
except if credits are required we either just show the credit amount if
a CreditCharge is set or show 'Incomplete' otherwise. Showing
'Incomplete' is equivalent to what we currently show to both types of
Users if a Case has not yet been completed in RT; this makes sense as
until an admin views the Cases dashboard we do not refresh the status of
RT tickets and so won't know whether a ticket is complete.

Various parts of this will probably change soon.