Replace usage of Octicons with Font Awesome
Font Awesome seems to be the standard for the new Flight web apps, so
might as well replace a few icons used from Octicons with this for
consistency.

Don't need to add Font Awesome as already included in
`app/assets/stylesheets/flight-branding/fonts.css`; also don't need to
add `font-awesome-rails` helper Gem as already included, but may as well
add to Gemfile to be explicit.