Ensure invalid AssetRecordFields not created
The previous technique of always initializing an AssetRecordField
associated with a Component or group, regardless of whether this would
ultimately be needed, was causing it to be possible to leave a Component
or group un-savable due to it having an association with an invalid
AssetRecordField (due to the nil value). Changing this to only
initialize an AssetRecordField when needed fixes this.

Also added assertions to ensure the Component/group is valid in each
situation (two of which fail without the corresponding code change).