Include all Case association info in single column
Rather than having a column for the Cluster associated with a Case, and
another for the associated Component, Service, or `N/A` if neither of
these is present, have a single column which always shows the Cluster as
well as the Component and Service if these are present.

This shows the same information but more clearly and concisely, and
should make it simpler to appropriately display a
Cluster/Service/Component as under maintenance due to a particular Case.