Handle associating other models with MaintenanceWindows
Previously the model a MaintenanceWindow was considered to be 'for' was
the associated model for the Case associated with it. However, now we
want to support putting a Cluster/Component/Service into maintenance in
association with a Case it was not originally associated with.

This commit supports this by allowing passing in a model to be
associated when using the RequestMaintenanceWindow service, and this
will then be appropriately associated with the created
MaintenanceWindow. If nothing is passed then the Case's associated model
will be used, which leads to similar behaviour as previously.