Use service from MaintenanceWindow form
So creating a MaintenanceWindow in this way happens in the same way as
creating a MaintenanceWindow for a specific Case already is.

Not bothering to handle error here atm as shouldn't be possible for this
to fail, and if it does this is admin-only so we'll just get an
exception which we can investigate.