Allow maintenance requests for Components/Cluster
In addition to allowing requesting maintenance in association with an
existing Case for Services, similarly support this for Components and
entire Clusters. This has not actually been requested for whole
Clusters, but it is trivial to add this as well so may as well do this
now.

In order to have this work needed to change RequestMaintenanceWindow to
use `present?` to check which ID it has been passed; since these can now
come from the form they may be empty strings, which are truthy but not
present.