Associate MW model iff none already associated
When creating a MaintenanceWindow using a factory, if it has no
associated model given (i.e. no associated Cluster, Component, or
Service) then associate a new Component with it so it is valid. If an
associated model is given then do not associate another one as this
would make it invalid.

This doesn't effect anything in existing code but will allow easily and
transparently creating MaintenanceWindows for things other than
Components.