Correct association between Cluster and MWs
Previously Clusters were associated with MaintenanceWindows via the
Cases for that Cluster. This made getting the open MaintenanceWindows
for the Cluster or any part of it straightforward, but made it harder to
get the MaintenanceWindows for just the Cluster as a whole, and is also
no longer correct now we store the associated model for a
MaintenanceWindow on the MaintenanceWindow itself.

This commit makes the necessary changes to break this relationship and
change it to use the correct relationships between Clusters and
MaintenanceWindows; in particular:

- `Cluster#open_maintenance_windows` functions the same as for
Components and Services, returning just the MaintenanceWindows
associated with this model directly.

- `Cluster#open_related_maintenance_windows` gives the open
MaintenanceWindows associated with the Cluster or any part of it, and is
now used when displaying MaintenanceWindows on the Cluster dashboard.

- The MaintenanceWindows controller now uses the correct associated
model for a MaintenanceWindow, rather than relying on the associated
Case's associated model (which may not be the same).