Show icons next to Cluster parts for all open MWs
For every open MaintenanceWindow associated with a
Cluster/Component/Service, show an appropriate icon alongside that
model's name when it is displayed in various places.

This is an improvement on what we did previously, as it distinguishes a
MaintenanceWindow which is awaiting confirmation from one which is
currently under maintenance, and also indicates the number of open
maintenance windows.