Simplify/correct showing Case MW creation link
With the new way that MaintenanceWindows work, it does not make sense to
show a single button to end these in the Cases table, as one Case may
have many MaintenanceWindows. Also the button to create a new
MaintenanceWindow for a Case and associated model should be shown iff
that model is not already under maintenance for that Case.