Remove Case maintenance indicators
These were not very useful, in addition to being inaccurate, now that a
Cluster or part can be under maintenance in association with any Case.
Removing them for now for this reason, which also greatly simplifies the
code; can always add similar but correct indicators back in future if
needed.