Automatically add Case Cluster if part given
This should make Cases a bit nicer to work with; if a Component or
Service is associated with a Case but no Cluster then the Case will now
automatically have the corresponding Cluster also associated, rather
than this just giving a validation error.

If a Cluster is explicitly associated with a Case then this will not be
replaced, and this will still give a validation error if it is invalid
given the associated Cluster part.