Add consultancy form for Clusters and parts
Similar to the new Case form this will create a support Case (using the
same `CasesController#create` action), but this page is specific to
consultancy issues. This means:

- the Cluster, Component, or Service to be associated is hard-coded
based on which object the form is accessed;

- the Issue to be associated is hard-coded to the appropriate
consultancy issue for the type of this object;

- only a details text area is shown, and the page content and
confirmation dialog makes clear that you may be charged for consultancy.