Do not render Site Case form button when no managed Clusters
The new Case form is now only applicable to managed Clusters, and only
the consultancy form is available to advice Clusters and their parts.
Therefore if a Site has no managed Clusters do not show the site-level
new Case form as there will be nothing useful to show on it.