Add simple scheme for checking read permissions
Add check whenever a model is loaded from the database, guaranteeing
that when a Site contact is logged in they can only access Site-specific
models that belong to their own Site; this should ensure we don't expose
any data from other Sites.

If such a permissions error does occur then it will be handled the same
as a 404 error, so we also don't expose whether such a resource exists
at all.