Refactor so Cases table can take any model
Rather than reference the Site and independently assigning an array of
Cases to be shown in the Cases table partial, just pass in a model and
read the Cases from this. This should make it possible to pass in
different types of models to display the Cases at different levels.