Remove CaseCategorys; add Categorys
This is the first step towards making Services the top-level model
selected when creating a Case. We want to remove the required segmenting
of Issues into CaseCategorys, which may have a controlling ServiceType,
and replace this with optional Categorys for Issues, and every Case will
always be associated with a Service (a change to be made in a future
commit).