Assign Case Cluster on initialize
Rather than just before validation. This is more correct, as we may
often want to access what will be the Cluster for a Case, without
actually having validated/saved it yet. In particular this fixes
accessing the Component/Service consultancy forms which were erroring
due to the lack of a Cluster being associated with the new Case at the
time the form is rendered.