Have Issues be nested under Services
In an effort to simplify the Case form we want Issues to be nested under
the Service they are available for, rather than stored independently
within CaseCategorys (which are not used at the moment, but will be
added back later as an optional extra grouping for ServiceTypes with
many Issues).

This change also simplifies the Case form app code in various places as
we no longer need to track which Issues and Services are compatible, and
there should later be other places we can simplify due to this change.

This doesn't completely work yet (in particular single Component mode
does not work at the moment, nor does creating Cases requiring
Components), but it compiles and works for Cluster and Service Cases.