Move data changes from obsolete migrations to seeds
Rather than trying to recreate these in new data migrations, which will
run in both the production server and locally every time the database is
seeded, just do this in the seed data as these data changes will already
have been made in production.

This works fine for now (locally at least), but probably longer term a
better way to handle this kind of globally available data would be to
just export the production data to ensure we are always in sync locally.