Simplify code ensuring Service-Issue compatibility
Every Issue which can be associated with a Case using the new Case form
now requires a Service, and the Ruby encoding of the data used to
initialize this form combined with the new State data model ensures only
a compatible Service and Issue can be selected, therefore we can remove
code to ensure the compatibility between these and only need to ensure
they have compatible support types.

This also fixes the creation of Cases for Issues requiring a Component,
as we now always encode the selected Service's ID, which is always
required.