Include Service Issues within categories
If any Issue which is applicable for a given Service has a Category,
include all the Issues for its Case form JSON nested within categories
(including an 'Other' Category for any which are uncategorised). If no
Issues are categorised then keep the same behaviour as before, since
adding a level of categories is unnecessary.