Update Case form app with optional categories
There can now be an optional categories level in the new Case form
between Services and the Issues which may be selected for each Service,
to allow further classifying of the available Issues where this is
desirable.

This commit updates the model to handle this new structure, as well as
appropriately handling decoding the Case form JSON into this and all
updates needed due to this new structure. Allowing changing the selected
Category has not yet been added.