Improve generated model permission methods tests
This commit improves the generated tests checking that every model has
the methods needed for permissions, by generating a test case for every
model (except the special case of Users) which checks that the model
either has a relation to a specific Site or is explicitly globally
available.

This should mean that one of these methods should be added to each
future model, helping us ensure that permissions are appropriately
handled. It also means we don't need to maintain a white-list of models
which are globally available in the tests, since this should already be
specified in the model itself.

As a consequence of this change I noticed that ComponentMake does not
specify one of these methods, so this has also been added.