Create (Default and Component)Expansion's using STI
With the `expansions` table as the base, Expanions can come in two
types: Defaut and Component. Please not this is independant of the
ExpansionType model which describes the physical type of the expansion.

Instead the `DefaultExpansion` model belongs_to a ComponentMake. It
will be used as part of the creation of new `Component`s. It however
does correspond with any physical hardware.

`ComponentExpansion` is related to a `Component` and represents a
peice of hardware that is sitting in a slot for a component.