Setup for CI with Travis
Add Travis config file for CI, and use setup script from this; this has
the dual benefit of sharing the development and testing setup, and
ensuring the setup script will be regularly run so it shouldn't get too
out-of-date.

The setup script has also been updated to include other necessary setup
steps.