Fix :site spec for (Component and Default)Expansion
This commit does not fix the spec for Expansion. Instead it sets the
DefaultExpansion's as global and ComponentExpansion's are associated
with a site.

However as Expansion is the base class for both of them it can not
be set either way. If it was, it would require either unsetting in
another class.

Instead this is a problem with the spec. I do not believe Expansion
should be forced either way and instead an exception in the spec
should probably be made.