Supress global site error for Expansion
The `Expansion` model is an abstact model that is used for STI
for both a global and "local" model (in regards to the site).

This means that `User` is not the only model with different behaviour.
Instead a whitelist of irregular classes as been created. This can
be updated in the future if STI is used extensively.