Create empty database in setup script
Rather than always seeding this with development data. This prevents
test failure in CI, due to unexpected data already existing in the
database, and will also avoid us blowing away a developer's existing
data when they run this script; if they want to reseed the database they
can do this independently of this script.