Enforce DefaultExpansion relationsip in the model
As the expansion table allows null values for the ComponentMake and
Component relationship, they need to be enforced in application logic.

In the case of `DefaultExpansion`, it must have a relationship to
`ComponentMake`. It does not have a `component` method, however it
will still respond to `component_id` which must be (null/ absent).