Create ComponentExpansion's based off defaults
When a new component is created, it will automatically create the
default expansions. This is done by reading the DefaultExpansions
from ComponentMake.

The issue with this approach is the :ports, :slot, and :expansion_type
attributes must be manually stated. Because the Expansions use STI,
not all the attributes are ported across.

In particular the id fields must not be copied as this could easily
make an invalid ComponentExpansion. Also the :type field must not be
copied as this will screw around with the STI.

I am open to suggestions on how to do this better. However for the
time being the list needs to be maintained. There is a test for this
behaviour so future changes should be detected.