Create concern for handling asset_record_fields
To dry the asset_record_fields in multiple models, a new concern has
been created. It will be based off the code for ComponentGroup.

The empty method will remain in ComponentGroup because it will need
to be redone in the future. It is not intended for it to become part
of the concern. (*may change)