Merge the asset_record_fields with it's parent class
Previous the asset_record_fields for a model where determined by what
was stored in the asset_record_layers method. This means that the
order has to be maintained in multiple modules and may get out of
sync.

Instead a new parent_asset_record_fields method has to be specified.
This needs to be an Enumerable object of the parent assets. The
new assets are then merged over the top of it similarly to before.

By chaining multiple models together, the same affect can be achieved
as before.