Switch ComponentGroup to use HasAssetRecordFields
The old code that was used to create the asset_record_fields has
been removed. Instead the parent_asset_record_fields is a modified
version of the empty_asset_record_fields.

Note the inclusion of the additional assets in asset_record_fields,
has broken the validation on the cluster. This will be fixed in a
latter commit.