Switch Component to use HasAssetRecordField
The code has been DRY'ed up and replaced with the Concern. The parent
class assets are from ComponentGroup which will access the blank
assets from ComponentType.

There is a bug caused by the collection being an array instead of
an ActiveRecord::...::Collection (of some description). This will
be fixed in a later commit.