Do not override the rails asset_record_field relation
The asset_record_field relation that is created by rails can not be
overridden with a simple array. This is because it is required to create
new asset_record_field for the component.

Instead a `combined_asset_record_field` method has been added. This
method will return the combined array instead. As the rails relation is
now valid again, the validation on asset_record_field has been included.

As it isn't obvious if `combined_asset_record_fields` or
`asset_record_fields` are required by `parent_asset_record_fields`, the
choice has been removed from the user. Instead the parent model itself
needs to be specified in `parent_for_asset_record_fields`. Then the
Concern will call the correct method (FYI it is the combined method).