Make asset_records the main focus of the concern
The combined method has been removed as it can be merged into the
asset_records method. The model no longer converts the array into a hash
as this limits the data available to the view.

Instead the view has been updated to use an array. ComponentType has
been updated to reflect the name change.