Remove hashify method from HasAssetRecords
The conversion into a hash is confusing and makes it difficult to
determine which values remain.

As a first pass to reduce the complexity, the merging of hashes has been
replaced with the injection of arrays.

A loop is still required (either in the form of a merge or inject)
however this removes the unnecessary data conversions.

Please note:
As the asset_record_fields is the starting memo, they have priority over
the parents records. The parent record is skipped if it has been updated
in the new id array (e.g. it is iun asset_record_fields).

This is currently an O(nm) { n: parent_records, m: new_records } that
needs to occur for each layer in the chain of asset_records. As the
layers are calculated concurrently, this makes the overall execution time
O(nml) { l: layer depth }.

It maybe possible to decouple the parent_records iteration O(n) from the
updating of the records O(nm) to make an overall O(n+ml) time. This
would require using a hash again where finding the parent_record to be
updated can be done O(1).

Thus all the records can be updated in O(ml) time for an arbitrary
layer. Then a single loop over the entire parent hash can be done in
O(n) time. Making the overall time O(ml+n).