Show edit form on AssetRecord page
The AssetRecordController has been updated to find the parameter id in a
more streamline fashion. The partial for the asset record has been
updated to show a form when in an edit scope.

Some additional code was required to determine which asset is currently
being edited which is tested in the spec.