Add submit button to AssetRecord and clean routes
The form for asset records has been updated to include a submit button.
This required a bit of reshuffling of the html, however this is only for
aesthetics.

The routes have been updated to include an update path for the asset
records. This has been done in a proc so the route doesn't need to be
maintained independently for both Component and ComponentGroup.

Please note the routing for the form is still broken, as it will be
submitted to *component-path*/asset_record/edit as a POST instead of
*component-path*/asset_record as a PATCH.

Due to the complexity of determining the route, this will need to be fix
in a latter commit.